add primapi unittest and remove some unused test cases.
