+++ python/paddle/fluid/tests/unittests/autograd/test_autograd_functional_dynamic.py
@@ -334,7 +334,7 @@ class TestJVP(TestAutogradFunctional):
     ('multi_in_single_out', paddle.matmul,
      (np.random.rand(2, 2), np.random.rand(2, 2))),
 ))
-class TestJacobianClassNoBatch(unittest.TestCase):
+class TestJacobianNoBatch(unittest.TestCase):
 
     def setUp(self):
         self._dtype = self.xs[0].dtype if isinstance(
@@ -387,7 +387,7 @@ class TestJacobianClassNoBatch(unittest.TestCase):
     ('3d_in_3d_out', utils.square, np.random.rand(2, 3, 4)),
     ('multi_in_single_out', utils.square, np.random.rand(2, 3)),
 ))
-class TestJacobianClassBatchFirst(unittest.TestCase):
+class TestJacobianBatchFirst(unittest.TestCase):
 
     def setUp(self):
         self._dtype = self.xs[0].dtype if isinstance(
@@ -444,7 +444,7 @@ class TestJacobianClassBatchFirst(unittest.TestCase):
         self.func_jacobian()
 
 
-class TestHessianClassNoBatch(unittest.TestCase):
+class TestHessianNoBatch(unittest.TestCase):
 
     @classmethod
     def setUpClass(self):
@@ -544,7 +544,7 @@ class TestHessianClassNoBatch(unittest.TestCase):
         self.func_out_not_single()
 
 
-class TestHessianClassBatchFirst(unittest.TestCase):
+class TestHessianBatchFirst(unittest.TestCase):
 
     @classmethod
     def setUpClass(self):
@@ -668,830 +668,6 @@ class TestHessianClassBatchFirst(unittest.TestCase):
         self.func_allow_unused()
         self.func_stop_gradient()
         self.func_out_not_single()
-
-
-class TestHessian(unittest.TestCase):
-
-    @classmethod
-    def setUpClass(self):
-        self.shape = (2, 2)
-        self.dtype = 'float32'
-        self.np_dtype = np.float32
-        self.numerical_delta = config.TOLERANCE.get(
-            self.dtype).get("second_order_grad").get("eps")
-        self.rtol = config.TOLERANCE.get(
-            self.dtype).get("second_order_grad").get("rtol")
-        self.atol = config.TOLERANCE.get(
-            self.dtype).get("second_order_grad").get("atol")
-
-        self.x = paddle.rand(shape=self.shape, dtype=self.dtype)
-        self.y = paddle.rand(shape=self.shape, dtype=self.dtype)
-
-    def func_single_input(self):
-
-        def func(x):
-            return paddle.sum(paddle.matmul(x, x))
-
-        numerical_hessian = _compute_numerical_hessian(func, self.x,
-                                                       self.numerical_delta,
-                                                       self.np_dtype)
-
-        self.x.stop_gradient = False
-        hessian = paddle.autograd.hessian(func, self.x)
-        np.testing.assert_allclose(hessian.numpy(), numerical_hessian[0][0],
-                                   self.rtol, self.atol)
-
-    def func_multi_input(self):
-
-        def func(x, y):
-            return paddle.sum(paddle.matmul(x, y))
-
-        numerical_hessian = _compute_numerical_hessian(func, [self.x, self.y],
-                                                       self.numerical_delta,
-                                                       self.np_dtype)
-
-        self.x.stop_gradient = False
-        self.y.stop_gradient = False
-        hessian = paddle.autograd.hessian(func, [self.x, self.y])
-        for i in range(len(hessian)):
-            for j in range(len(hessian[0])):
-                np.testing.assert_allclose(hessian[i][j].numpy(),
-                                           numerical_hessian[i][j], self.rtol,
-                                           self.atol)
-
-    def func_allow_unused_false(self):
-
-        def func(x, y):
-            return paddle.sum(paddle.matmul(x, x))
-
-        try:
-            self.x.stop_gradient = False
-            self.y.stop_gradient = False
-            hessian = paddle.autograd.hessian(func, [self.x, self.y])
-        except ValueError as e:
-            error_msg = cpt.get_exception_message(e)
-            assert error_msg.find("allow_unused") > 0
-
-    def func_allow_unused_true(self):
-
-        def func(x, y):
-            return paddle.sum(paddle.matmul(x, x))
-
-        numerical_hessian = _compute_numerical_hessian(func, [self.x, self.y],
-                                                       self.numerical_delta,
-                                                       self.np_dtype)
-        self.x.stop_gradient = False
-        self.y.stop_gradient = False
-        hessian = paddle.autograd.hessian(func, [self.x, self.y],
-                                          allow_unused=True)
-        for i in range(len(hessian)):
-            for j in range(len(hessian[0])):
-                if i == j == 0:
-                    np.testing.assert_allclose(hessian[i][j].numpy(),
-                                               numerical_hessian[i][j],
-                                               self.rtol, self.atol)
-                else:
-                    assert hessian[i][j] is None
-
-    def func_create_graph_false(self):
-
-        def func(x):
-            return paddle.sum(paddle.matmul(x, x))
-
-        numerical_hessian = _compute_numerical_hessian(func, self.x,
-                                                       self.numerical_delta,
-                                                       self.np_dtype)
-        self.x.stop_gradient = False
-        hessian = paddle.autograd.hessian(func, self.x)
-        assert hessian.stop_gradient == True
-        np.testing.assert_allclose(hessian.numpy(), numerical_hessian[0][0],
-                                   self.rtol, self.atol)
-        try:
-            paddle.grad(hessian, self.x)
-        except Exception as e:
-            error_msg = cpt.get_exception_message(e)
-            assert error_msg.find("has no gradient") > 0 or error_msg.find(
-                "does not appear") > 0
-
-    def func_create_graph_true(self):
-        fluid.set_flags({"FLAGS_retain_grad_for_all_tensor": True})
-
-        def func(x):
-            return paddle.sum(F.sigmoid(x))
-
-        numerical_hessian = _compute_numerical_hessian(func, self.x,
-                                                       self.numerical_delta,
-                                                       self.np_dtype)
-        self.x.stop_gradient = False
-        hessian = paddle.autograd.hessian(func, self.x, create_graph=True)
-        assert hessian.stop_gradient == False
-        np.testing.assert_allclose(hessian.numpy(), numerical_hessian[0][0],
-                                   self.rtol, self.atol)
-        triple_grad = paddle.grad(hessian, self.x)
-        assert triple_grad is not None
-        fluid.set_flags({"FLAGS_retain_grad_for_all_tensor": False})
-
-    def test_all_cases(self):
-        with _test_eager_guard():
-            self.setUpClass()
-            self.func_single_input()
-            self.func_multi_input()
-            self.func_allow_unused_false()
-            self.func_allow_unused_true()
-            self.func_create_graph_false()
-            self.func_create_graph_true()
-        self.setUpClass()
-        self.func_single_input()
-        self.func_multi_input()
-        self.func_allow_unused_false()
-        self.func_allow_unused_true()
-        self.func_create_graph_false()
-        self.func_create_graph_true()
-
-
-class TestHessianFloat64(TestHessian):
-
-    @classmethod
-    def setUpClass(self):
-        self.shape = (2, 2)
-        self.dtype = 'float64'
-        self.np_dtype = np.float64
-        self.numerical_delta = config.TOLERANCE.get(
-            self.dtype).get("second_order_grad").get("eps")
-        self.rtol = config.TOLERANCE.get(
-            self.dtype).get("second_order_grad").get("rtol")
-        self.atol = config.TOLERANCE.get(
-            self.dtype).get("second_order_grad").get("atol")
-        self.x = paddle.rand(shape=self.shape, dtype=self.dtype)
-        self.y = paddle.rand(shape=self.shape, dtype=self.dtype)
-
-
-class TestBatchHessian(unittest.TestCase):
-
-    @classmethod
-    def setUpClass(self):
-        self.x_shape = (5, 2)
-        self.weight_shape = (2, 4)
-        self.y_shape = (5, 2)
-        self.dtype = 'float32'
-        self.np_dtype = np.float32
-        self.numerical_delta = config.TOLERANCE.get(
-            self.dtype).get("second_order_grad").get("eps")
-        self.rtol = config.TOLERANCE.get(
-            self.dtype).get("second_order_grad").get("rtol")
-        self.atol = config.TOLERANCE.get(
-            self.dtype).get("second_order_grad").get("atol")
-        self.x = paddle.rand(shape=self.x_shape, dtype=self.dtype)
-        self.weight = paddle.rand(shape=self.weight_shape, dtype=self.dtype)
-        self.y = paddle.rand(shape=self.y_shape, dtype=self.dtype)
-
-    def func_single_input(self):
-
-        def func(x):
-            return paddle.matmul(x * x, self.weight)[:, 0:1]
-
-        numerical_hessian = _compute_numerical_batch_hessian(
-            func, self.x, self.numerical_delta, self.np_dtype)
-        self.x.stop_gradient = False
-        hessian = paddle.autograd.batch_hessian(func, self.x, create_graph=True)
-        np.testing.assert_allclose(hessian, numerical_hessian, self.rtol,
-                                   self.atol)
-
-    def func_multi_input(self):
-
-        def func(x, y):
-            return paddle.matmul(x * x * y * y, self.weight)[:, 0:1]
-
-        numerical_hessian = _compute_numerical_batch_hessian(
-            func, [self.x, self.y], self.numerical_delta, self.np_dtype)
-
-        self.x.stop_gradient = False
-        self.y.stop_gradient = False
-        hessian = paddle.autograd.batch_hessian(func, [self.x, self.y])
-
-        shape_tensor = paddle.to_tensor(numerical_hessian).astype("float64")
-        hessian_reshape = np.reshape(hessian, (shape_tensor.shape))
-        np.testing.assert_allclose(hessian_reshape, numerical_hessian,
-                                   self.rtol, self.atol)
-
-    def func_allow_unused_false(self):
-
-        def func(x, y):
-            return paddle.matmul(x * x, self.weight)[:, 0:1]
-
-        try:
-            self.x.stop_gradient = False
-            self.y.stop_gradient = False
-            hessian = paddle.autograd.batch_hessian(func, [self.x, self.y])
-        except ValueError as e:
-            error_msg = cpt.get_exception_message(e)
-            assert error_msg.find("allow_unused") > 0
-
-    def func_allow_unused_true(self):
-
-        def func(x, y):
-            return paddle.matmul(x * x, self.weight)[:, 0:1]
-
-        numerical_hessian = _compute_numerical_batch_hessian(
-            func, [self.x, self.y], self.numerical_delta, self.np_dtype)
-        self.x.stop_gradient = False
-        self.y.stop_gradient = False
-        hessian = paddle.autograd.batch_hessian(func, [self.x, self.y],
-                                                allow_unused=True)
-
-        for i in range(len(hessian)):
-            for j in range(len(hessian[0])):
-                if i == j == 0:
-                    numerical_hessian = np.stack(
-                        (numerical_hessian[i][j], numerical_hessian[i][j + 1]),
-                        axis=0)
-                    np.testing.assert_allclose(hessian[i][j], numerical_hessian,
-                                               self.rtol, self.atol)
-                else:
-                    assert hessian[i][j] is None
-
-    def func_create_graph_false(self):
-
-        def func(x):
-            return paddle.matmul(x * x, self.weight)[:, 0:1]
-
-        numerical_hessian = _compute_numerical_batch_hessian(
-            func, self.x, self.numerical_delta, self.np_dtype)
-        self.x.stop_gradient = False
-        hessian = paddle.autograd.batch_hessian(func, self.x)
-        assert hessian.stop_gradient == True
-        np.testing.assert_allclose(hessian.numpy(), numerical_hessian,
-                                   self.rtol, self.atol)
-        try:
-            paddle.grad(hessian, self.x)
-        except Exception as e:
-            error_msg = cpt.get_exception_message(e)
-            assert error_msg.find("has no gradient") > 0 or error_msg.find(
-                "does not appear") > 0
-
-    def func_create_graph_true(self):
-
-        def func(x):
-            return paddle.matmul(x * x, self.weight)[:, 0:1]
-
-        numerical_hessian = _compute_numerical_batch_hessian(
-            func, self.x, self.numerical_delta, self.np_dtype)
-        self.x.stop_gradient = False
-        hessian = paddle.autograd.batch_hessian(func, self.x, create_graph=True)
-        assert hessian.stop_gradient == False
-        np.testing.assert_allclose(hessian.numpy(), numerical_hessian,
-                                   self.rtol, self.atol)
-        triple_grad = paddle.grad(hessian, self.x)
-        assert triple_grad is not None
-
-    def test_all_cases(self):
-        with _test_eager_guard():
-            self.setUpClass()
-            self.func_single_input()
-            self.func_multi_input()
-            self.func_allow_unused_false()
-            self.func_allow_unused_true()
-            self.func_create_graph_false()
-            self.func_create_graph_true()
-        self.setUpClass()
-        self.func_single_input()
-        self.func_multi_input()
-        self.func_allow_unused_false()
-        self.func_allow_unused_true()
-        self.func_create_graph_false()
-        self.func_create_graph_true()
-
-
-class TestBatchHessianFloat64(TestBatchHessian):
-
-    @classmethod
-    def setUpClass(self):
-        self.x_shape = (5, 2)
-        self.weight_shape = (2, 4)
-        self.y_shape = (5, 2)
-        self.dtype = 'float64'
-        self.np_dtype = np.float64
-        self.numerical_delta = config.TOLERANCE.get(
-            self.dtype).get("second_order_grad").get("eps")
-        self.rtol = config.TOLERANCE.get(
-            self.dtype).get("second_order_grad").get("rtol")
-        self.atol = config.TOLERANCE.get(
-            self.dtype).get("second_order_grad").get("atol")
-        self.x = paddle.rand(shape=self.x_shape, dtype=self.dtype)
-        self.weight = paddle.rand(shape=self.weight_shape, dtype=self.dtype)
-        self.y = paddle.rand(shape=self.y_shape, dtype=self.dtype)
-
-
-class TestVHP(unittest.TestCase):
-
-    @classmethod
-    def setUpClass(self):
-        self.shape = (2, 2)
-        self.dtype = 'float32'
-        self.np_dtype = np.float32
-        self.numerical_delta = config.TOLERANCE.get(
-            self.dtype).get("second_order_grad").get("eps")
-        self.rtol = config.TOLERANCE.get(
-            self.dtype).get("second_order_grad").get("rtol")
-        self.atol = config.TOLERANCE.get(
-            self.dtype).get("second_order_grad").get("atol")
-        self.x = paddle.rand(shape=self.shape, dtype=self.dtype)
-        self.y = paddle.rand(shape=self.shape, dtype=self.dtype)
-        self.vx = paddle.rand(shape=self.shape, dtype=self.dtype)
-        self.vy = paddle.rand(shape=self.shape, dtype=self.dtype)
-
-    def func_single_input(self):
-
-        def func(x):
-            return paddle.sum(paddle.matmul(x, x))
-
-        numerical_func_output = func(self.x).numpy()
-        numerical_vhp = _compute_numerical_vhp(func, self.x, self.vx,
-                                               self.numerical_delta,
-                                               self.np_dtype)
-
-        self.x.stop_gradient = False
-        func_output, vhp = paddle.autograd.vhp(func, self.x, self.vx)
-        np.testing.assert_allclose(func_output.numpy(), numerical_func_output,
-                                   self.rtol, self.atol)
-        np.testing.assert_allclose(vhp[0].numpy(), numerical_vhp[0], self.rtol,
-                                   self.atol)
-
-    def func_multi_input(self):
-
-        def func(x, y):
-            return paddle.sum(paddle.matmul(x, y))
-
-        numerical_func_output = func(self.x, self.y).numpy()
-        numerical_vhp = _compute_numerical_vhp(func, [self.x, self.y],
-                                               [self.vx, self.vy],
-                                               self.numerical_delta,
-                                               self.np_dtype)
-
-        self.x.stop_gradient = False
-        self.y.stop_gradient = False
-        func_output, vhp = paddle.autograd.vhp(func, [self.x, self.y],
-                                               [self.vx, self.vy])
-        np.testing.assert_allclose(func_output.numpy(), numerical_func_output,
-                                   self.rtol, self.atol)
-        for i in range(len(vhp)):
-            np.testing.assert_allclose(vhp[i].numpy(), numerical_vhp[i],
-                                       self.rtol, self.atol)
-
-    def func_v_default(self):
-
-        def func(x, y):
-            return paddle.sum(paddle.matmul(x, y))
-
-        numerical_func_output = func(self.x, self.y).numpy()
-        vx = paddle.ones(self.vx.shape, dtype=self.vx.dtype)
-        vy = paddle.ones(self.vy.shape, dtype=self.vy.dtype)
-        numerical_vhp = _compute_numerical_vhp(func, [self.x, self.y], [vx, vy],
-                                               self.numerical_delta,
-                                               self.np_dtype)
-
-        self.x.stop_gradient = False
-        self.y.stop_gradient = False
-        func_output, vhp = paddle.autograd.vhp(func, [self.x, self.y])
-        np.testing.assert_allclose(func_output.numpy(), numerical_func_output,
-                                   self.rtol, self.atol)
-        for i in range(len(vhp)):
-            np.testing.assert_allclose(vhp[i].numpy(), numerical_vhp[i],
-                                       self.rtol, self.atol)
-
-    def func_allow_unused_true(self):
-
-        def func(x, y):
-            return paddle.sum(paddle.matmul(x, x))
-
-        numerical_func_output = func(self.x, self.y).numpy()
-        numerical_vhp = _compute_numerical_vhp(func, [self.x, self.y],
-                                               [self.vx, self.vy],
-                                               self.numerical_delta,
-                                               self.np_dtype)
-
-        self.x.stop_gradient = False
-        self.y.stop_gradient = False
-        func_output, vhp = paddle.autograd.vhp(func, [self.x, self.y],
-                                               [self.vx, self.vy])
-        np.testing.assert_allclose(func_output.numpy(), numerical_func_output,
-                                   self.rtol, self.atol)
-        np.testing.assert_allclose(vhp[0].numpy(), numerical_vhp[0], self.rtol,
-                                   self.atol)
-
-    def func_create_graph_true(self):
-        fluid.set_flags({"FLAGS_retain_grad_for_all_tensor": True})
-
-        def func(x):
-            return paddle.sum(F.sigmoid(x))
-
-        numerical_func_output = func(self.x).numpy()
-        numerical_vhp = _compute_numerical_vhp(func, self.x, self.vx,
-                                               self.numerical_delta,
-                                               self.np_dtype)
-
-        self.x.stop_gradient = False
-        func_output, vhp = paddle.autograd.vhp(func, self.x, self.vx)
-        np.testing.assert_allclose(func_output.numpy(), numerical_func_output,
-                                   self.rtol, self.atol)
-        assert vhp[0].stop_gradient == False
-        np.testing.assert_allclose(vhp[0].numpy(), numerical_vhp[0], self.rtol,
-                                   self.atol)
-        triple_grad = paddle.grad(vhp, self.x)
-        assert triple_grad is not None
-        fluid.set_flags({"FLAGS_retain_grad_for_all_tensor": False})
-
-    def test_all_cases(self):
-        with _test_eager_guard():
-            self.setUpClass()
-            self.func_v_default()
-            self.func_multi_input()
-            self.func_single_input()
-            self.func_allow_unused_true()
-            self.func_create_graph_true()
-        self.setUpClass()
-        self.func_v_default()
-        self.func_multi_input()
-        self.func_single_input()
-        self.func_allow_unused_true()
-        self.func_create_graph_true()
-
-
-class TestJacobian(unittest.TestCase):
-
-    @classmethod
-    def setUpClass(self):
-        self.shape = (4, 4)
-        self.dtype = 'float32'
-        self.np_dtype = np.float32
-        self.numerical_delta = 1e-4
-        self.rtol = 1e-3
-        self.atol = 1e-3
-        self.x = paddle.rand(shape=self.shape, dtype=self.dtype)
-        self.y = paddle.rand(shape=self.shape, dtype=self.dtype)
-
-    def func_single_input_and_single_output(self):
-
-        def func(x):
-            return paddle.matmul(x, x)
-
-        numerical_jacobian = _compute_numerical_jacobian(
-            func, self.x, self.numerical_delta, self.np_dtype)
-        self.x.stop_gradient = False
-        jacobian = paddle.autograd.jacobian(func, self.x)
-        np.testing.assert_allclose(jacobian.numpy(), numerical_jacobian[0][0],
-                                   self.rtol, self.atol)
-
-    def func_single_input_and_multi_output(self):
-
-        def func(x):
-            return paddle.matmul(x, x), x * x
-
-        numerical_jacobian = _compute_numerical_jacobian(
-            func, self.x, self.numerical_delta, self.np_dtype)
-        self.x.stop_gradient = False
-        jacobian = paddle.autograd.jacobian(func, self.x)
-        for i in range(len(jacobian)):
-            np.testing.assert_allclose(jacobian[i].numpy(),
-                                       numerical_jacobian[i][0], self.rtol,
-                                       self.atol)
-
-    def func_multi_input_and_single_output(self):
-
-        def func(x, y):
-            return paddle.matmul(x, y)
-
-        numerical_jacobian = _compute_numerical_jacobian(
-            func, [self.x, self.y], self.numerical_delta, self.np_dtype)
-        self.x.stop_gradient = False
-        self.y.stop_gradient = False
-        jacobian = paddle.autograd.jacobian(func, [self.x, self.y])
-        for j in range(len(jacobian)):
-            np.testing.assert_allclose(jacobian[j].numpy(),
-                                       numerical_jacobian[0][j], self.rtol,
-                                       self.atol)
-
-    def func_multi_input_and_multi_output(self):
-        fluid.set_flags({"FLAGS_retain_grad_for_all_tensor": True})
-
-        def func(x, y):
-            return paddle.matmul(x, y), x * y
-
-        numerical_jacobian = _compute_numerical_jacobian(
-            func, [self.x, self.y], self.numerical_delta, self.np_dtype)
-        self.x.stop_gradient = False
-        self.y.stop_gradient = False
-        jacobian = paddle.autograd.jacobian(func, [self.x, self.y])
-        for i in range(len(jacobian)):
-            for j in range(len(jacobian[0])):
-                np.testing.assert_allclose(jacobian[i][j].numpy(),
-                                           numerical_jacobian[i][j], self.rtol,
-                                           self.atol)
-        fluid.set_flags({"FLAGS_retain_grad_for_all_tensor": False})
-
-    def func_allow_unused_false(self):
-
-        def func(x, y):
-            return paddle.matmul(x, x)
-
-        try:
-            self.x.stop_gradient = False
-            self.y.stop_gradient = False
-            jacobian = paddle.autograd.jacobian(func, [self.x, self.y])
-        except ValueError as e:
-            error_msg = cpt.get_exception_message(e)
-            assert error_msg.find("allow_unused") > 0
-
-    def func_allow_unused_true(self):
-
-        def func(x, y):
-            return paddle.matmul(x, x)
-
-        numerical_jacobian = _compute_numerical_jacobian(
-            func, [self.x, self.y], self.numerical_delta, self.np_dtype)
-        self.x.stop_gradient = False
-        self.y.stop_gradient = False
-        jacobian = paddle.autograd.jacobian(func, [self.x, self.y],
-                                            allow_unused=True)
-        np.testing.assert_allclose(jacobian[0].numpy(),
-                                   numerical_jacobian[0][0], self.rtol,
-                                   self.atol)
-        assert jacobian[1] is None
-
-    def func_create_graph_false(self):
-
-        def func(x, y):
-            return paddle.matmul(x, y)
-
-        numerical_jacobian = _compute_numerical_jacobian(
-            func, [self.x, self.y], self.numerical_delta, self.np_dtype)
-        self.x.stop_gradient = False
-        self.y.stop_gradient = False
-        jacobian = paddle.autograd.jacobian(func, [self.x, self.y])
-        for j in range(len(jacobian)):
-            assert jacobian[j].stop_gradient == True
-            np.testing.assert_allclose(jacobian[j].numpy(),
-                                       numerical_jacobian[0][j], self.rtol,
-                                       self.atol)
-        try:
-            paddle.grad(jacobian[0], [self.x, self.y])
-        except Exception as e:
-            error_msg = cpt.get_exception_message(e)
-            assert error_msg.find("has no gradient") > 0 or error_msg.find(
-                "does not appear") > 0
-
-    def func_create_graph_true(self):
-
-        def func(x, y):
-            return paddle.matmul(x, y)
-
-        numerical_jacobian = _compute_numerical_jacobian(
-            func, [self.x, self.y], self.numerical_delta, self.np_dtype)
-        self.x.stop_gradient = False
-        self.y.stop_gradient = False
-        jacobian = paddle.autograd.jacobian(func, [self.x, self.y],
-                                            create_graph=True)
-        for j in range(len(jacobian)):
-            assert jacobian[j].stop_gradient == False
-            np.testing.assert_allclose(jacobian[j].numpy(),
-                                       numerical_jacobian[0][j], self.rtol,
-                                       self.atol)
-        double_grad = paddle.grad(jacobian[0], [self.x, self.y])
-        assert double_grad is not None
-
-    def test_all_cases(self):
-        with _test_eager_guard():
-            self.setUpClass()
-            self.func_multi_input_and_multi_output()
-            self.func_multi_input_and_single_output()
-            self.func_single_input_and_multi_output()
-            self.func_single_input_and_single_output()
-            self.func_allow_unused_false()
-            self.func_allow_unused_true()
-            self.func_create_graph_false()
-            self.func_create_graph_true()
-        self.setUpClass()
-        self.func_multi_input_and_multi_output()
-        self.func_multi_input_and_single_output()
-        self.func_single_input_and_multi_output()
-        self.func_single_input_and_single_output()
-        self.func_allow_unused_false()
-        self.func_allow_unused_true()
-        self.func_create_graph_false()
-        self.func_create_graph_true()
-
-
-class TestJacobianFloat64(TestJacobian):
-
-    @classmethod
-    def setUpClass(self):
-        self.shape = (4, 4)
-        self.dtype = 'float64'
-        self.np_dtype = np.float64
-        self.numerical_delta = 1e-7
-        self.rtol = 1e-7
-        self.atol = 1e-7
-        self.x = paddle.rand(shape=self.shape, dtype=self.dtype)
-        self.y = paddle.rand(shape=self.shape, dtype=self.dtype)
-
-
-class TestJacobianBatch(unittest.TestCase):
-
-    @classmethod
-    def setUpClass(self):
-        self.x_shape = (4, 2)
-        self.weight_shape = (2, 4)
-        self.y_shape = (4, 2)
-        self.dtype = 'float32'
-        self.np_dtype = np.float32
-        self.numerical_delta = 1e-4
-        self.rtol = 1e-3
-        self.atol = 1e-3
-        self.x = paddle.rand(shape=self.x_shape, dtype=self.dtype)
-        self.weight = paddle.rand(shape=self.weight_shape, dtype=self.dtype)
-        self.y = paddle.rand(shape=self.y_shape, dtype=self.dtype)
-
-    def func_batch_single_input_and_batch_single_output(self):
-
-        def func(x):
-            return paddle.matmul(paddle.matmul(x, self.weight), self.y)
-
-        numerical_jacobian = _compute_numerical_batch_jacobian(
-            func, [self.x], self.numerical_delta, self.np_dtype)
-
-        self.x.stop_gradient = False
-        batch_jacobian = paddle.autograd.batch_jacobian(
-            func,
-            self.x,
-        )
-
-        self.assertTrue(
-            np.allclose(batch_jacobian.numpy().all(),
-                        numerical_jacobian[0][0].all()))
-
-    def func_batch_single_input_and_batch_multi_output(self):
-
-        def func(x):
-            return paddle.matmul(paddle.matmul(x, self.weight), self.y), x * x
-
-        numerical_jacobian = _compute_numerical_batch_jacobian(
-            func, [self.x], self.numerical_delta, self.np_dtype)
-
-        self.x.stop_gradient = False
-        batch_jacobian = paddle.autograd.batch_jacobian(
-            func,
-            self.x,
-        )
-
-        for i in range(len(batch_jacobian)):
-            np.testing.assert_allclose(batch_jacobian[i].numpy(),
-                                       numerical_jacobian[i][0], self.rtol,
-                                       self.atol)
-
-    def func_batch_multi_input_and_batch_single_output(self):
-
-        def func(x, y):
-            return x * y
-
-        numerical_jacobian = _compute_numerical_batch_jacobian(
-            func, [self.x, self.y], self.numerical_delta, self.np_dtype)
-
-        self.x.stop_gradient = False
-        self.y.stop_gradient = False
-        batch_jacobian = paddle.autograd.batch_jacobian(func, [self.x, self.y])
-
-        for j in range(len(batch_jacobian)):
-            np.testing.assert_allclose(batch_jacobian[j].numpy(),
-                                       numerical_jacobian[0][j], self.rtol,
-                                       self.atol)
-
-    def func_batch_multi_input_and_batch_multi_output(self):
-
-        def func(x, y):
-            return x * y, x * y
-
-        numerical_jacobian = _compute_numerical_batch_jacobian(
-            func, [self.x, self.y], self.numerical_delta, self.np_dtype)
-
-        self.x.stop_gradient = False
-        self.y.stop_gradient = False
-        batch_jacobian = paddle.autograd.batch_jacobian(func, [self.x, self.y])
-
-        for i in range(len(batch_jacobian)):
-            np.testing.assert_allclose(batch_jacobian[i], numerical_jacobian[i],
-                                       self.rtol, self.atol)
-
-    def func_allow_unused_false(self):
-
-        def func(x, y):
-            return x * x
-
-        try:
-            self.x.stop_gradient = False
-            self.y.stop_gradient = False
-            jacobian = paddle.autograd.batch_jacobian(func, [self.x, self.y])
-        except ValueError as e:
-            error_msg = cpt.get_exception_message(e)
-            assert error_msg.find("allow_unused") > 0
-
-    def func_allow_unused_true(self):
-
-        def func(x, y):
-            return x * x
-
-        numerical_jacobian = _compute_numerical_batch_jacobian(
-            func, [self.x, self.y], self.numerical_delta, self.np_dtype)
-        self.x.stop_gradient = False
-        self.y.stop_gradient = False
-        jacobian = paddle.autograd.batch_jacobian(func, [self.x, self.y],
-                                                  allow_unused=True)
-
-        np.testing.assert_allclose(jacobian[0].numpy(),
-                                   numerical_jacobian[0][0], self.rtol,
-                                   self.atol)
-        assert jacobian[1] is None
-
-    def func_create_graph_false(self):
-
-        def func(x, y):
-            return x * y
-
-        numerical_jacobian = _compute_numerical_batch_jacobian(
-            func, [self.x, self.y], self.numerical_delta, self.np_dtype)
-        self.x.stop_gradient = False
-        self.y.stop_gradient = False
-        jacobian = paddle.autograd.batch_jacobian(func, [self.x, self.y])
-        for j in range(len(jacobian)):
-            assert jacobian[j].stop_gradient == True
-            np.testing.assert_allclose(jacobian[j].numpy(),
-                                       numerical_jacobian[0][j], self.rtol,
-                                       self.atol)
-        try:
-            paddle.grad(jacobian[0], [self.x, self.y])
-        except Exception as e:
-            error_msg = cpt.get_exception_message(e)
-            assert error_msg.find("has no gradient") > 0 or error_msg.find(
-                "does not appear") > 0
-
-    def func_create_graph_true(self):
-
-        def func(x, y):
-            return x * y
-
-        numerical_jacobian = _compute_numerical_batch_jacobian(
-            func, [self.x, self.y], self.numerical_delta, self.np_dtype)
-        self.x.stop_gradient = False
-        self.y.stop_gradient = False
-        jacobian = paddle.autograd.batch_jacobian(func, [self.x, self.y],
-                                                  create_graph=True)
-        for j in range(len(jacobian)):
-            assert jacobian[j].stop_gradient == False
-            np.testing.assert_allclose(jacobian[j].numpy(),
-                                       numerical_jacobian[0][j], self.rtol,
-                                       self.atol)
-        double_grad = paddle.grad(jacobian[0], [self.x, self.y])
-        assert double_grad is not None
-
-    def test_all_cases(self):
-        with _test_eager_guard():
-            self.setUpClass()
-            self.func_batch_single_input_and_batch_single_output()
-            self.func_batch_single_input_and_batch_multi_output()
-            self.func_batch_multi_input_and_batch_single_output()
-            self.func_batch_multi_input_and_batch_multi_output()
-            self.func_allow_unused_false()
-            self.func_allow_unused_true()
-            self.func_create_graph_false()
-            self.func_create_graph_true()
-        self.setUpClass()
-        self.func_batch_single_input_and_batch_single_output()
-        self.func_batch_single_input_and_batch_multi_output()
-        self.func_batch_multi_input_and_batch_single_output()
-        self.func_batch_multi_input_and_batch_multi_output()
-        self.func_allow_unused_false()
-        self.func_allow_unused_true()
-        self.func_create_graph_false()
-        self.func_create_graph_true()
-
-
-class TestJacobianBatchFloat64(TestJacobianBatch):
-
-    @classmethod
-    def setUpClass(self):
-        self.x_shape = (12, 2)
-        self.weight_shape = (2, 12)
-        self.y_shape = (12, 2)
-        self.dtype = 'float64'
-        self.np_dtype = np.float64
-        self.numerical_delta = config.TOLERANCE.get(
-            self.dtype).get('second_order_grad').get('eps')
-        self.rtol = config.TOLERANCE.get(
-            self.dtype).get('second_order_grad').get('rtol')
-        self.atol = config.TOLERANCE.get(
-            self.dtype).get('second_order_grad').get('atol')
-        self.x = paddle.rand(shape=self.x_shape, dtype=self.dtype)
-        self.weight = paddle.rand(shape=self.weight_shape, dtype=self.dtype)
-        self.y = paddle.rand(shape=self.y_shape, dtype=self.dtype)
 
 
 if __name__ == "__main__":
